+++ src/components/Input/Input.js
@@ -2,11 +2,13 @@
  * @Author: wangweixin@threatbook.cn
  * @Date: 2018-01-18 17:50:52
  * @Last Modified by: wangweixin@threatbook.cn
- * @Last Modified time: 2018-08-21 15:14:39
+ * @Last Modified time: 2018-08-21 17:45:07
  */
 import React, { PureComponent } from 'react'
 import classNames from 'classnames'
 import ControledInput from '../Common/ControledInput'
+import Icon from '../Icon'
+import searchIcon from '@/images/svg/search.svg'
 
 @ControledInput(
   v => v,
@@ -14,7 +16,7 @@ import ControledInput from '../Common/ControledInput'
 )
 export default class Input extends PureComponent {
   render() {
-    const { type, hasError, className, props: controled, style, ...others } = this.props
+    const { type, isSearch, hasError, className, props: controled, ...others } = this.props
     const classes = classNames(
       'input',
       {
@@ -22,6 +24,12 @@ export default class Input extends PureComponent {
       },
       className
     )
-    return <input className={classes} type={type} style={style} {...controled} {...others} />
+    return <div className="input-wrapper">
+      <input className={classes} placeholder={isSearch ? '搜索相关内容' : ''} type={type} {...others} {...controled} />
+      { isSearch
+        ? <Icon className="search-icon" link={searchIcon}/>
+        : ''
+      }
+    </div>
   }
 }